Add files via upload
This picture shows the icons I would like to have. One is Twitter (it must link to the website twitter account). The other is called "Seeking Alpha", the icon must link to the website's Seeking Alpha account, here is the website http://seekingalpha.com/.  This is why I was asking for a website email, so I can make the 2 accounts and link them to website.

At the bottom of the homepage "Content copyright 2017. Deliverance Capital. All rights reserved. "